Fixed the Sample Rate
Had an issue raised to me on Nexus Mods. Turns out the sample rate was failing if it wasn't "16-bit 44k". Pushed a fix to make checks for this and added in support for .ogg and .mp3 to be converted into the correct .wav format as well to add more ease of use to the app.

Also added some more information to message boxes to make things more clear as to where files are being saved.