V1.2/V1.3 Fixes for WAV and Admin
V1.2 Fixes: Turns out if the .WAV doesn't follow the 16-bit 44k format then VGAudioCli runs into some issues. Fixed it to just convert the audio to the format required to avoid any issues with weird conversions. Also figured if we're auto converting it anyway I might as well add support in for .MP3 and .OGG.

V1.3 Fixes: Turns out there's a Windows security feature that disables drag and drop functions for apps if you're running an app as an admin, which sadly was the main thing I had the app doing. I added in a button that opens a File Explorer box, this lets the user choose their files instead. I could've done some PInvoke stuff to bypass it but that felt like a lot of extra permissions that could potentially lead to the app having permission issues when it gets launched. I decided to just set up a warning message when the app detects it's in Admin mode and lets the user know that Drag and Drop is disabled but they can still use the button to add songs in. This will be handy for some reason they can't get out of Admin mode.